Move CacheStats to its own file.
