+++ invokeai/backend/model_manager/load/model_cache/__init__.py
@@ -1,6 +1,6 @@
 """Init file for ModelCache."""
 
-from .model_cache_base import ModelCacheBase, CacheStats  # noqa F401
+from .model_cache_base import ModelCacheBase  # noqa F401
 from .model_cache_default import ModelCache  # noqa F401
 
-_all__ = ["ModelCacheBase", "ModelCache", "CacheStats"]
+_all__ = ["ModelCacheBase", "ModelCache"]
